backend/requirements.txt
fastapi
uvicorn
openai
langchain
python-dotenv
langchain-community
langchain_openai